# docs/public/openapi/regesta-v0.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Regesta V0 HTTP API",
    "version": "0.0.0",
    "summary": "Machine-readable reference for the current Regesta V0 HTTP API surface.",
    "description": "This document describes implemented core registry routes and the current npm projection routes. Regesta-native object shapes are referenced from the V0 JSON Schema reference. Protocol semantics remain defined by the TypeScript protocol and core validation code."
  },
  "servers": [
    {
      "url": "https://registry.example",
      "description": "Core registry API host"
    },
    {
      "url": "https://npm.registry.example",
      "description": "npm projection host"
    }
  ],
...
  "paths": {
...
    "/-/package/{encoded}/dist-tags": {
      "get": {
        "tags": ["npm Projection"],
        "summary": "Read npm dist-tags for an encoded package name",
        "parameters": [{ "$ref": "#/components/parameters/EncodedNpmName" }],
        "responses": {
          "200": {
            "description": "npm dist-tags object.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NpmDistTags" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" },
          "502": { "$ref": "#/components/responses/Error" }
        }
      },
      "head": {
        "tags": ["npm Projection"],
        "summary": "Read npm dist-tags headers for an encoded package name",
        "parameters": [{ "$ref": "#/components/parameters/EncodedNpmName" }],
        "responses": {
          "200": { "description": "dist-tags exist." },
          "404": { "$ref": "#/components/responses/Error" },
          "502": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/-/package/{scope}/{name}/dist-tags": {
      "get": {
        "tags": ["npm Projection"],
        "summary": "Read npm dist-tags for a scoped package name",
        "parameters": [
          { "$ref": "#/components/parameters/NpmScope" },
          { "$ref": "#/components/parameters/NpmName" }
        ],
        "responses": {
          "200": {
            "description": "npm dist-tags object.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NpmDistTags" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" },
          "502": { "$ref": "#/components/responses/Error" }
        }
      },
      "head": {
        "tags": ["npm Projection"],
        "summary": "Read npm dist-tags headers for a scoped package name",
        "parameters": [
          { "$ref": "#/components/parameters/NpmScope" },
          { "$ref": "#/components/parameters/NpmName" }
        ],
        "responses": {
          "200": { "description": "dist-tags exist." },
          "404": { "$ref": "#/components/responses/Error" },
          "502": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/{name}/-/{file}": {
      "get": {
        "tags": ["npm Projection"],
        "summary": "Read npm tarball for an unscoped package",
        "parameters": [
          { "$ref": "#/components/parameters/NpmName" },
          { "$ref": "#/components/parameters/NpmTarballFile" }
        ],
        "responses": {
          "200": {
            "description": "npm install artifact bytes.",
            "content": {
              "application/gzip": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "206": { "description": "npm tarball byte range." },
          "304": { "$ref": "#/components/responses/NotModified" },
          "404": { "$ref": "#/components/responses/Error" },
          "416": { "description": "Requested range not satisfiable." }
        }
      },
      "head": {
        "tags": ["npm Projection"],
        "summary": "Read npm tarball headers for an unscoped package",
        "parameters": [
          { "$ref": "#/components/parameters/NpmName" },
          { "$ref": "#/components/parameters/NpmTarballFile" }
        ],
        "responses": {
          "200": { "description": "npm tarball exists." },
          "304": { "$ref": "#/components/responses/NotModified" },
          "404": { "$ref": "#/components/responses/Error" },
          "416": { "description": "Requested range not satisfiable." }
        }
      }
    },
    "/{scope}/{name}/-/{file}": {
      "get": {
        "tags": ["npm Projection"],
        "summary": "Read npm tarball for a scoped package",
        "parameters": [
          { "$ref": "#/components/parameters/NpmScope" },
          { "$ref": "#/components/parameters/NpmName" },
          { "$ref": "#/components/parameters/NpmTarballFile" }
        ],
        "responses": {
          "200": {
            "description": "npm install artifact bytes.",
            "content": {
              "application/gzip": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "206": { "description": "npm tarball byte range." },
          "304": { "$ref": "#/components/responses/NotModified" },
          "404": { "$ref": "#/components/responses/Error" },
          "416": { "description": "Requested range not satisfiable." }
        }
      },
      "head": {
        "tags": ["npm Projection"],
        "summary": "Read npm tarball headers for a scoped package",
        "parameters": [
          { "$ref": "#/components/parameters/NpmScope" },
          { "$ref": "#/components/parameters/NpmName" },
          { "$ref": "#/components/parameters/NpmTarballFile" }
        ],
        "responses": {
          "200": { "description": "npm tarball exists." },
          "304": { "$ref": "#/components/responses/NotModified" },
          "404": { "$ref": "#/components/responses/Error" },
          "416": { "description": "Requested range not satisfiable." }
        }
      }
    }
  },
  "components": {
...
    "parameters": {
...
      "EncodedNpmName": {
        "in": "path",
        "name": "encoded",
        "required": true,
        "schema": { "$ref": "#/components/schemas/NonEmptyString" }
      },
...
      "NpmName": {
        "in": "path",
        "name": "name",
        "required": true,
        "schema": { "$ref": "#/components/schemas/NonEmptyString" }
      },
      "NpmScope": {
        "in": "path",
        "name": "scope",
        "required": true,
        "schema": { "$ref": "#/components/schemas/NonEmptyString" }
      },
...
      "NpmTarballFile": {
        "in": "path",
        "name": "file",
        "required": true,
        "schema": { "$ref": "#/components/schemas/NonEmptyString" }
      },
...
    },
    "responses": {
      "Error": {
        "description": "Structured public error response.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotModified": {
        "description": "Client validator still matches current representation."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "code": { "type": "string" },
          "error": { "type": "string" },
          "message": { "type": "string" }
        },
        "required": ["error"]
      },
...
      "NonEmptyString": {
        "type": "string",
        "minLength": 1
      },
      "NpmDistTags": {
        "type": "object",
        "additionalProperties": { "type": "string" }
      },
...
    }
  }
}